Move device manager to native_inference
