+++ src/marqo/inference/native_inference/local_inference.py
@@ -1,6 +1,6 @@
 from marqo.core.exceptions import DeviceError
-from marqo.core.inference.device_manager import DeviceManager
 import marqo.core.inference.api.exceptions as inference_api_exceptions
+from marqo.inference.native_inference.device_manager import DeviceManager
 from marqo.inference.native_inference.embedding_models.hugging_face_model import HuggingFaceModel
 from marqo.inference.native_inference.embedding_models.open_clip_model import OpenCLIPModel
 from marqo.inference.native_inference.embedding_models.random_model import RandomModel
